feat(user): 외부에서도 user 패키지 호출 가능하도록 수정
