+++ src/main/kotlin/com/fastcampus/commerce/user/api/controller/UserController.kt
@@ -5,51 +5,65 @@ import com.fastcampus.commerce.auth.api.dto.RegisterRequest
 import com.fastcampus.commerce.auth.interfaces.web.security.oauth.NaverUserResponse
 import com.fastcampus.commerce.user.api.dto.UserDto
 import com.fastcampus.commerce.user.api.service.UserService
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@RequestMapping("/api/v1/user")
+@RequestMapping("/api/user")
 class UserController(
     private val userService: UserService,
 ) {
-    fun registerUser(request: RegisterRequest): UserDto {
-        return UserDto(
-            id = "",
-            externalId = "",
-            name = "",
-            email = request.email,
-            nickname = "User",
-            profileImage = "",
-            roles = emptyList(),
-        )
+
+    /**
+     * Registers a new user.
+     *
+     * @param registerRequest the registration request containing user details
+     * @return the registered user data
+     */
+    @PostMapping("/register")
+    @ResponseBody
+    fun registerUser(@RequestBody registerRequest: RegisterRequest): UserDto {
+        return userService.registerUser(registerRequest)
     }
 
-    fun loginUser(request: LoginRequest): UserDto {
-        return UserDto(
-            id = "",
-            externalId = "",
-            name = "",
-            email = request.email,
-            nickname = "User",
-            profileImage = "",
-            roles = emptyList(),
-        )
+    /**
+     * Logs in a user.
+     *
+     * @param loginRequest the login request containing credentials
+     * @return the logged-in user data
+     */
+    @PostMapping("/login")
+    @ResponseBody
+    fun loginUser(@RequestBody loginRequest: LoginRequest): UserDto {
+        return userService.loginUser(loginRequest)
     }
 
-    fun deleteUser(userId: Long) {
-        // TODO: delete 처리
+    /**
+     * Deletes a user by user ID.
+     *
+     * @param userId the ID of the user to delete
+     */
+    @DeleteMapping("/delete/{userId}")
+    fun deleteUser(@PathVariable userId: Long) {
+        userService.deleteUser(userId)
     }
 
-    fun saveOrUpdateUser(naverUser: NaverUserResponse): UserDto {
-        return UserDto(
-            id = "",
-            externalId = "",
-            name = "",
-            email = naverUser.email,
-            nickname = "User",
-            profileImage = "",
-            roles = emptyList(),
-        )
+    /**
+     * Saves or updates a user using Naver user response data.
+     *
+     * @param naverUserResponse the Naver user response data
+     * @return the saved or updated user data
+     */
+    @PutMapping("/naver")
+    @ResponseBody
+    fun saveOrUpdateUser(@RequestBody naverUserResponse: NaverUserResponse): UserDto {
+        return userService.saveOrUpdateUser(naverUserResponse)
     }
+
 }
